+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- json extension to URL path, e.g. /api/v1/users.json, /api/v1/users/luke.json
+- Error messages in body if requests fails
+- Validate user input in request bodies with validator
+- Requests with queries, e.g. /permissions?uname=luke;gname=d1-read
 ### Changed
 ### Deprecated
 ### Removed